+++ src/components/hotels/HotelCard/HotelCard.styles.jsx
