+++ src/components/hotels/HotelMapView.jsx
