+++ src/components/hotels/HotelRating.jsx
