+++ src/components/hotels/PhotoGallery.jsx
