+++ src/components/hotels/RatingBadge.jsx
