+++ src/components/shared/PhotoGallery.jsx
