+++ src/pages/HotelDetailsPage.jsx
