+++ src/pages/HotelsPage.jsx
